requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
Babel==2.8.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
cloudinary==1.19.1
colorama==0.4.3
coverage==5.0.3
dj-database-url==0.5.0
Django==3.0.2
django-cors-headers==3.2.1
django-countries==5.5
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-filter==2.2.0
django-jenkins==0.110.0
django-ordered-model==3.3.0
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.7
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
mock==3.0.5
pep8==1.7.1
phonenumbers==8.11.1
Pillow==7.0.0
psycopg2==2.8.4
pycryptodome==3.9.4
pyflakes3k==0.4.3
PyJWT==1.7.1
pylint==2.4.4
python-http-client==3.2.4
python-stdnum==1.13
pytz==2019.3
rave-python==1.1.2
requests==2.22.0
sendgrid==6.1.0
six==1.13.0
sqlparse==0.3.0
suds-jurko==0.6
typed-ast==1.4.1
urllib3==1.25.7
vatnumber==1.2
vine==1.3.0
whitenoise==5.0.1
wrapt==1.11.2
zipp==2.2.0
